Move dom definitions to react-instantsearch-dom
