+++ types/react-instantsearch/react-instantsearch-tests.tsx
@@ -1,280 +1,9 @@
 import * as React from "react";
-import { InstantSearch, Hits, Highlight, SearchBox, RefinementList, CurrentRefinements, ClearRefinements, Pagination, Menu, Configure, createConnector, Index } from 'react-instantsearch/dom';
-import { connectHighlight, connectSearchBox, connectMenu, connectRefinementList, connectStateResults, SearchBoxProvided, Hit, SearchState } from "react-instantsearch/connectors";
+
+import { connectMenu, connectRefinementList, connectStateResults, SearchState } from "react-instantsearch/connectors";
+import { InstantSearch, SearchBox, Index, Hits, Highlight, Menu } from "react-instantsearch/dom";
 import { orderBy, omit, values } from 'lodash';
 import { createInstantSearch } from "react-instantsearch-core";
-
-// DOM
-// https://community.algolia.com/react-instantsearch/Getting_started.html
-() => {
-  const App = () => (
-    <InstantSearch
-      appId="latency"
-      apiKey="3d9875e51fbd20c7754e65422f7ce5e1"
-      indexName="bestbuy"
-    >
-      <Search />
-    </InstantSearch>
-  );
-
-  function Search() {
-    return (
-      <div className="container">
-        <Hits />
-      </div>
-    );
-  }
-
-  function Product({ hit }: { hit: Hit }) {
-    return <div>{hit.name}</div>;
-  }
-
-  function Search2() {
-    return (
-      <div>
-        <Hits hitComponent={Product} />
-      </div>
-    );
-  }
-
-  function Product2({ hit }: { hit: Hit }) {
-    return (
-      <div style={{ marginTop: '10px' }}>
-        <span className="hit-name">
-          <Highlight attribute="name" hit={hit} />
-        </span>
-      </div>
-    );
-  }
-
-  function Search3() {
-    return (
-      <div className="container">
-        <SearchBox />
-        <RefinementList attribute="category" />
-        <Hits hitComponent={Product} />
-      </div>
-    );
-  }
-
-  function Search4() {
-    return (
-      <div className="container">
-        <CurrentRefinements />
-        <ClearRefinements />
-        <SearchBox />
-        <RefinementList attribute="category" />
-        <Hits hitComponent={Product} />
-        <Pagination />
-      </div>
-    );
-  }
-};
-
-// https://community.algolia.com/react-instantsearch/guide//Highlighting_results.html
-() => {
-  const Hit = ({ hit }: { hit: Hit }) => (
-    <p>
-      <Highlight attribute="description" hit={hit} tagName="mark" />
-    </p>
-  );
-
-  function App() {
-    return (
-      <InstantSearch
-        appId="latency"
-        apiKey="6be0576ff61c053d5f9a3225e2a90f76"
-        indexName="instant_search"
-      >
-        <Hits hitComponent={Hit} />
-      </InstantSearch>
-    );
-  }
-};
-
-// TODO
-// () => {
-//   const CustomHighlight = connectHighlight(
-//     ({ highlight, attribute, hit, highlightProperty }) => {
-//       const parsedHit = highlight({
-//         attribute,
-//         hit,
-//         highlightProperty: '_highlightResult'
-//       });
-//       const highlightedHits = parsedHit.map(part => {
-//         if (part.isHighlighted) return <mark>{part.value}</mark>;
-//         return part.value;
-//       });
-//       return <div>{highlightedHits}</div>;
-//     }
-//   );
-
-//   const Hit = ({ hit }: { hit: Hit }) => (
-//     <p>
-//       <CustomHighlight attribute="description" hit={hit} />
-//     </p>
-//   );
-
-//   function App() {
-//     return (
-//       <InstantSearch
-//         appId="latency"
-//         apiKey="6be0576ff61c053d5f9a3225e2a90f76"
-//         indexName="instant_search"
-//       >
-//         <Hits hitComponent={Hit} />
-//       </InstantSearch>
-//     );
-//   }
-// };
-
-// https://community.algolia.com/react-instantsearch/guide/i18n.html
-() => {
-  const App = () => (
-    <InstantSearch
-      appId="..."
-      apiKey="..."
-      indexName="..."
-    >
-      <Menu
-        attribute="fruits"
-        translations={{ showMore: 'Voir plus' }}
-      />
-    </InstantSearch>
-  );
-};
-
-// https://community.algolia.com/react-instantsearch/guide/Sorting_and_filtering.html
-// TODO
-// () => {
-//   const App = () => (
-//     <InstantSearch
-//       appId="..."
-//       apiKey="..."
-//       indexName="..."
-//     >
-//       <SearchBox defaultRefinement="hi" />
-//       <RefinementList
-//         attribute="category"
-//         transformItems={items =>
-//           orderBy(items, ['label', 'count'], ['asc', 'desc'])
-//         }
-//       />
-//     </InstantSearch>
-//   );
-// };
-
-// https://community.algolia.com/react-instantsearch/guide/Connectors.html
-() => {
-  const MySearchBox = ({currentRefinement, refine}: SearchBoxProvided) =>
-    <input
-      type="text"
-      value={currentRefinement}
-      onChange={e => refine(e.target.value)}
-    />;
-
-  // `ConnectedSearchBox` renders a `<MySearchBox>` widget that is connected to
-  // the <InstantSearch> state, providing it with `currentRefinement` and `refine` props for
-  // reading and manipulating the current query of the search.
-  const ConnectedSearchBox = connectSearchBox(MySearchBox);
-};
-
-// https://community.algolia.com/react-instantsearch/guide/Default_refinements.html
-() => {
-  const App = () => (
-    <InstantSearch
-      appId="..."
-      apiKey="..."
-      indexName="..."
-    >
-      <SearchBox defaultRefinement="hi" />
-      <Menu attribute="fruits" defaultRefinement="Orange" />
-    </InstantSearch>
-  );
-};
-
-// TODO
-// () => {
-//   const VirtualMenu = connectMenu(() => null);
-
-//   const App = () => (
-//     <InstantSearch
-//       appId="..."
-//       apiKey="..."
-//       indexName="..."
-//     >
-//       <div>
-//         <CurrentRefinements
-//           transformItems={items =>
-//             items.filter(item => item.currentRefinement !== 'Orange')
-//           }
-//         />
-//         <SearchBox/>
-//         <VirtualMenu attribute="fruits" defaultRefinement="Orange" />
-//         <Menu attribute="origin" defaultRefinement="Spain" />
-//       </div>
-//     </InstantSearch>
-//   );
-// };
-
-// https://community.algolia.com/react-instantsearch/guide/Searching_in_Lists.html
-
-<RefinementList attribute="products" searchable />;
-
-() => {
-  const RefinementListWithSearchBox = connectRefinementList(props => {
-    const values = props.items.map(item => {
-      const label = item._highlightResult
-        ? <Highlight attribute="label" hit={item}/>
-        : item.label;
-
-      return (
-        <li key={item.value}>
-          <span onClick={() => props.refine(item.value)}>
-            {label} {item.isRefined ? '- selected' : ''}
-          </span>
-        </li>
-      );
-    });
-
-    return (
-      <div>
-        <input type="search" onInput={e => props.searchForItems((e.target as HTMLInputElement).value)}/>
-        <ul>{values}</ul>
-      </div>
-    );
-  });
-
-  return <RefinementListWithSearchBox attribute="products" />;
-};
-
-// https://community.algolia.com/react-instantsearch/guide/Virtual_widgets.html
-() => {
-  const VirtualMenu = connectMenu(() => null);
-  const Hoodies = () => <VirtualMenu attribute="clothes" defaultRefinement="hoodies" />;
-
-  const App = () => (
-    <InstantSearch
-      appId="..."
-      apiKey="..."
-      indexName="..."
-    >
-      <SearchBox defaultRefinement="hi" />
-      <Hoodies />
-      <Menu attribute="fruits" defaultRefinement="Orange" />
-    </InstantSearch>
-  );
-};
-
-// https://community.algolia.com/react-instantsearch/guide/Search_parameters.html
-<InstantSearch
-  appId="appId"
-  apiKey="apiKey"
-  indexName="indexName"
->
-  <Configure distinct={1}/>
-  // widgets
-</InstantSearch>;
 
 // https://community.algolia.com/react-instantsearch/guide/Search_state.html
 () => {
@@ -609,4 +338,49 @@ import { createInstantSearch } from "react-instantsearch-core";
   );
 
   <InstantSearch />;
+};
+
+() => {
+  const RefinementListWithSearchBox = connectRefinementList(props => {
+    const values = props.items.map(item => {
+      const label = item._highlightResult
+        ? <Highlight attribute="label" hit={item}/>
+        : item.label;
+
+      return (
+        <li key={item.value}>
+          <span onClick={() => props.refine(item.value)}>
+            {label} {item.isRefined ? '- selected' : ''}
+          </span>
+        </li>
+      );
+    });
+
+    return (
+      <div>
+        <input type="search" onInput={e => props.searchForItems((e.target as HTMLInputElement).value)}/>
+        <ul>{values}</ul>
+      </div>
+    );
+  });
+
+  return <RefinementListWithSearchBox attribute="products" />;
+};
+
+// https://community.algolia.com/react-instantsearch/guide/Virtual_widgets.html
+() => {
+  const VirtualMenu = connectMenu(() => null);
+  const Hoodies = () => <VirtualMenu attribute="clothes" defaultRefinement="hoodies" />;
+
+  const App = () => (
+    <InstantSearch
+      appId="..."
+      apiKey="..."
+      indexName="..."
+    >
+      <SearchBox defaultRefinement="hi" />
+      <Hoodies />
+      <Menu attribute="fruits" defaultRefinement="Orange" />
+    </InstantSearch>
+  );
 };
